requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
telebot==0.0.5
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
yarl==1.20.1